nginx-all-practice/analytics-dashboard/py/requirements.txt
flask
psycopg2
requests
python-dotenv 